Move scanners in sub module
