+++ keycloak_scanner/scanners/form_post_xss_scanner.py
@@ -1,5 +1,5 @@
-from .scanner import Scanner
-from .custom_logging import verbose, find
+from keycloak_scanner.scanners.scanner import Scanner
+from keycloak_scanner.custom_logging import verbose, find
