+++ keycloak_scanner/scanners/well_known_scanner.py
@@ -1,6 +1,6 @@
 from keycloak_scanner.custom_logging import verbose, info
 from keycloak_scanner.properties import add_kv
-from keycloak_scanner.scanner import Scanner
+from keycloak_scanner.scanners.scanner import Scanner
 
 URL_PATTERN = '{}/auth/realms/{}/.well-known/openid-configuration'
